core/kompress: move vendorized to _deprecated, use kompress library directly
