+++ my/kython/kompress.py
@@ -1,1 +1,6 @@
-../core/kompress.py
+from my.core import __NOT_HPI_MODULE__
+from my.core import warnings
+
+warnings.high('my.kython.kompress is deprecated, please use "kompress" library directly. See https://github.com/karlicoss/kompress')
+
+from my.core.kompress import *
